Threw almost everything out and using blindbit-lib
The old structure of the frontend hindered further development and had
to be removed before it became crippling in future. It was a nice
attempt for quick prototypting but started becoming a problem.

Using BlindBit Lib now. It has most of the wallet and scanning stuff
standardised. BlindBit Desktop should be mainly the GUI and not the
underlying wallet scanning logic - Separation of concerns