Clean up settings page, rm buggy network change
Changing networks creates several issues. Wallet data will be overridden
and would either need to be stored in a separate location
(subdirectories) or needs to be handled in some other special way.
Changing the network normally also includes changing birthheight and
last sync height would need to be changed as well. Several factors to
consider. Removing GUI feature until a proper process is created for it.
Manual changes to the wallet.dat are always possible as long as the file
is not encrypted